Flytt contexter til der de gir mening
