+++ apps/selvbetjening-ui/src/main.tsx
@@ -5,19 +5,22 @@ import { fetcher } from './utils/api.ts'
 import { SWRConfig } from 'swr'
 import { ProvideSpraakContext } from './common/spraak/SpraakContext.tsx'
 import { InntektsjusteringRoot } from './inntektsjustering/InntektjusteringRoot.tsx'
+import { ProvideInnloggetInnbyggerContext } from './common/innloggetInnbygger/InnloggetInnbyggerContext.tsx'
 
 createRoot(document.getElementById('root')!).render(
     <StrictMode>
-        <ProvideSpraakContext>
-            <SWRConfig
-                value={{
-                    fetcher: fetcher,
-                    shouldRetryOnError: false,
-                    revalidateOnFocus: false,
-                }}
-            >
-                <InntektsjusteringRoot />
-            </SWRConfig>
-        </ProvideSpraakContext>
+        <SWRConfig
+            value={{
+                fetcher: fetcher,
+                shouldRetryOnError: false,
+                revalidateOnFocus: false,
+            }}
+        >
+            <ProvideSpraakContext>
+                <ProvideInnloggetInnbyggerContext>
+                    <InntektsjusteringRoot />
+                </ProvideInnloggetInnbyggerContext>
+            </ProvideSpraakContext>
+        </SWRConfig>
     </StrictMode>
 )
